Changed token_id type from int to bigint in transactions and transactions_meme_lab

+++ src/entities/ITransaction.ts
@@ -1,4 +1,4 @@
-import { Entity, Column, PrimaryColumn, Index } from 'typeorm';
+import { Column, Entity, Index, PrimaryColumn } from 'typeorm';
 
 export class BaseTransaction {
   @Column({ type: 'datetime' })
@@ -27,7 +27,7 @@ export class BaseTransaction {
   contract!: string;
 
   @Index()
-  @PrimaryColumn({ type: 'int' })
+  @PrimaryColumn({ type: 'bigint' })
   token_id!: number;
 
   @Column({ type: 'int' })
